Unpack OCI images.
The process uses `*at` syscalls (like `linkat` or `openat2`) in all filesystem
actions to ensure that files are never written outside the target directory.

The `unpack` function is replaced with a builder (`Unpacker::new`), so it will
be easier to add more parameters without breaking the API.

The umask is disabled during the process to ensure that the new files are
created with the expected mode. The code assumes that no other threads are
modifying the umask during `unpack`.

File ownership (uid/gid) is applied only if it is a non-zero value. Errors from
fchownat are ignored, so it can be used rootless.

The rules for the sandbox now use ABI 2 to be compatible with Linux 6.1.

The public API is now documented.